Create smart car parking system
It’s more crucial than ever to park your automobile efficiently in the fast-paced world of today. With the development of technology, we can now come up with clever methods to deal with this problem. In this project, we are going to study the construction of an Arduino-based car parking system with IR sensors. With the aid of this technology, parking your automobile will be simple and perfectly aligned within the parking place. We will guide you through the parts and procedures needed to build this ground-breaking system. MArobotics develops a smart “Car Parking System” that uses Arduino and infrared sensors to optimize parking management. This novel invention uses infrared (IR) sensors to identify the presence of automobiles in parking lots. Using Arduino’s capabilities, the system interprets sensor data to assess parking availability and sends real-time updates to users. The MArobotics Car Parking System not only streamlines the parking procedure but also improves efficiency by reducing search time for available spots, resulting in a better and more structured approach to urban parking difficulties.